Move calculateCheckDigit to new utility file
